# Cddo.Data.Marketplace.UI/Pages/Support/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CDDO Utilities",
    "version": "v1"
  },
  "paths": {
...
    "/Topics": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "parameters": [
          {
            "name": "DataAssetStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.DataAssets.Enums.DataAssetStatus"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoTopicsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoTopicsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoTopicsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Organisations": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "parameters": [
          {
            "name": "DataAssetStatuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.DataAssets.Enums.DataAssetStatus"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoOrganisationsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoOrganisationsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoOrganisationsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/query-catalog-reports-data": {
      "post": {
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/download-catalog-reports-data": {
      "post": {
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField": {
        "enum": [
          "ProfileId",
          "DataAssetType",
          "DataAssetId",
          "Title",
          "AlternativeTitles",
          "Summary",
          "Description",
          "Keywords",
          "Themes",
          "ContactPointName",
          "ContactPointEmailAddress",
          "Creator",
          "Publisher",
          "SecurityClassification",
          "Identifier",
          "RelatedResource",
          "Version",
          "Issued",
          "Modified",
          "Created",
          "UpdateFrequency",
          "Licence",
          "AccessRights",
          "DistributionTitle",
          "DistributionAccessService",
          "DistributionIdentifier",
          "DistributionModified",
          "DistributionIssued",
          "DistributionLicence",
          "DistributionByteSize",
          "DistributionMediaType",
          "DistributionDownloadUrl",
          "ServiceType",
          "ServiceStatus",
          "EndpointUrl",
          "EndpointDescription",
          "ServesData",
          "DataAssetStatus",
          "IsOwnerPointOfContact",
          "OwnerName",
          "OwnerEmail",
          "ApiType",
          "HasPart"
        ],
        "type": "string"
      },
...
      "Cddo.Data.Marketplace.Api.Dto.Models.DataAssets.Enums.DataAssetStatus": {
        "enum": [
          "Draft",
          "Published",
          "Archived",
          "Deleted",
          "Withdrawn"
        ],
        "type": "string"
      },
...
      "Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.CatalogReportsDataItem": {
        "type": "object",
        "properties": {
          "catalogReportsDataItemFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.CatalogReportsDataItemField"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.CatalogReportsDataItemField": {
        "required": [
          "field",
          "values"
        ],
        "type": "object",
        "properties": {
          "field": {
            "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.Filters.CatalogReportFieldFilter": {
        "type": "object",
        "properties": {
          "field": {
            "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.Filters.CatalogReportsFilter": {
        "type": "object",
        "properties": {
          "filterByInitiatingUserPermissions": {
            "type": "boolean"
          },
          "fieldFilters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.Filters.CatalogReportFieldFilter"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Cddo.Data.Marketplace.Api.Dto.Requests.Reports.QueryCatalogReportsDataRequest": {
        "required": [
          "numberOfRecords",
          "requiredFields",
          "startRecordIndex"
        ],
        "type": "object",
        "properties": {
          "requiredFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField"
            }
          },
          "filter": {
            "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.Filters.CatalogReportsFilter"
          },
          "startRecordIndex": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfRecords": {
            "type": "integer",
            "format": "int32"
          },
          "searchText": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoOrganisationsResponse": {
        "required": [
          "organisations"
        ],
        "type": "object",
        "properties": {
          "organisations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Cddo.Data.Marketplace.Api.Dto.Responses.Lookup.GetCddoTopicsResponse": {
        "required": [
          "topics"
        ],
        "type": "object",
        "properties": {
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Cddo.Data.Marketplace.Api.Dto.Responses.Reports.QueryCatalogReportsDataResponse": {
        "type": "object",
        "properties": {
          "totalNumberOfMatchedRecords": {
            "type": "integer",
            "format": "int32"
          },
          "selectableFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField"
            },
            "nullable": true
          },
          "selectedFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.CatalogData.CatalogAssetField"
            },
            "nullable": true
          },
          "selectedStatuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.DataAssets.Enums.DataAssetStatus"
            },
            "nullable": true
          },
          "organisations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "selectableOrganisations": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "reportName": {
            "type": "string",
            "nullable": true
          },
          "templateId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "catalogReportsDataItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cddo.Data.Marketplace.Api.Dto.Models.Reports.CatalogData.CatalogReportsDataItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}